Alex remote commit

+++ client_code/components/GridView.py
@@ -413,7 +413,7 @@ class GridView:
             db_row = self.grid_class.get(grid_row.uid) if grid_row.uid else None
             if db_row is not None:
                 db_row.delete()
-            self.grid.deleteRecord('uid', grid_row)
+            self.grid.deleteRecord('uid', grid_row) if grid_row.uid else self.grid.deleteRecord()
         if self.confirm_dialog:
             self.confirm_dialog.hide()
             self.confirm_dialog.destroy()
